multiboot2_common::test_utils

Struct AlignedBytes

#[repr(C, align(8))]
pub struct AlignedBytes<const N: usize>(pub [u8; N]);

Helper to 8-byte align the underlying bytes, as mandated in the Multiboot2 spec.

With this type, one can create manual and raw Multiboot2 boot information or just the bytes for simple tags, in a manual and raw approach.

Tuple Fields
0: [u8; N]
Implementations

impl<const N: usize> AlignedBytes<N>

pub const fn new(bytes: [u8; N]) -> Self

Creates a new type.